refactor: Move ArrowedTextButton for common use (#173)
- Move `ArrowedTextButton` from `presentation.rentaldetail.components` to the shared `common.component` package.
- Update import paths for `ArrowedTextButton` in various rental detail content screens.
